refactor: global.security → domain.security
