+++ dongmanee/src/main/java/com/dongmanee/domain/security/config/CorsConfig.java
@@ -1,4 +1,4 @@
-package com.dongmanee.global.security.config;
+package com.dongmanee.domain.security.config;
 
 import java.util.List;
