+++ dongmanee/src/main/java/com/dongmanee/domain/security/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.dongmanee.global.security.config;
+package com.dongmanee.domain.security.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -13,10 +13,10 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 
-import com.dongmanee.global.security.login.entrypoint.CustomAuthenticationEntryPoint;
-import com.dongmanee.global.security.login.filter.JwtAuthenticationFilter;
-import com.dongmanee.global.security.login.handler.CustomAccessDeniedHandler;
-import com.dongmanee.global.security.login.provider.JwtProvider;
+import com.dongmanee.domain.security.entrypoint.CustomAuthenticationEntryPoint;
+import com.dongmanee.domain.security.filter.JwtAuthenticationFilter;
+import com.dongmanee.domain.security.handler.CustomAccessDeniedHandler;
+import com.dongmanee.domain.security.provider.JwtProvider;
 
 import lombok.RequiredArgsConstructor;
 
@@ -29,15 +29,13 @@ public class SecurityConfig {
 
 	@Bean
 	public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-		http
-			.httpBasic(AbstractHttpConfigurer::disable) // spring security 기본 인증 해제
+		http.httpBasic(AbstractHttpConfigurer::disable) // spring security 기본 인증 해제
 			.formLogin(AbstractHttpConfigurer::disable)
 			.csrf(AbstractHttpConfigurer::disable) // CSRF 공격 방지 기능 해제
 			.cors(Customizer.withDefaults()) // CORS 설정
 
 			.headers((headerConfig) ->    // h2-console 표시를 위해
-				headerConfig.frameOptions(HeadersConfigurer.FrameOptionsConfig::disable)
-			)
+				headerConfig.frameOptions(HeadersConfigurer.FrameOptionsConfig::disable))
 			.authorizeHttpRequests(authorizedRequests -> { // HTTP 요청에 대한 인가 설정
 				authorizedRequests.anyRequest().permitAll(); // 모든 요청에 대해서 인증 없이 허용
 			})
