+++ dongmanee/src/main/java/com/dongmanee/domain/security/dto/request/SignInRequest.java
@@ -1,4 +1,4 @@
-package com.dongmanee.global.security.login.dto.request;
+package com.dongmanee.domain.security.dto.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
@@ -13,7 +13,6 @@ public class SignInRequest {
 	private String loginId;
 
 	@NotBlank(message = "비밀번호를 입력해주세요")
-	@Pattern(regexp = "(?=.*[0-9])(?=.*[a-zA-Z])(?=.*\\W)(?=\\S+$).{8,16}",
-		message = "비밀번호는 8~16자 영문 대 소문자, 숫자, 특수문자를 사용하세요.")
+	@Pattern(regexp = "(?=.*[0-9])(?=.*[a-zA-Z])(?=.*\\W)(?=\\S+$).{8,16}", message = "비밀번호는 8~16자 영문 대 소문자, 숫자, 특수문자를 사용하세요.")
 	private String password;
 }
