+++ dongmanee/src/main/java/com/dongmanee/domain/security/service/UserDetailLoginService.java
@@ -1,0 +1,26 @@
+package com.dongmanee.domain.security.service;
+
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+import com.dongmanee.domain.member.dao.MemberRepository;
+import com.dongmanee.domain.member.domain.Member;
+import com.dongmanee.domain.security.domain.MemberDetails;
+import com.dongmanee.domain.security.exception.UnAuthorizedException;
+
+import lombok.RequiredArgsConstructor;
+
+@Service
+@RequiredArgsConstructor
+public class UserDetailLoginService implements UserDetailsService {
+	private final MemberRepository memberRepository;
+
+	@Override
+	public UserDetails loadUserByUsername(String userId) throws UsernameNotFoundException {
+		Member member = memberRepository.findByLoginId(userId).orElseThrow(UnAuthorizedException::new);
+
+		return MemberDetails.of(member);
+	}
+}
